Add a new local testing package and updated the refactored wasm tests to use it
Added a new local "testing" package to abstract common elements used in
the refactored wasm tests. This package must only be used for testing.

Currently it contains 3 sub packages:

chromedp - for all of the headless chromedp functionality. The functionality is unchanged from the original code.
pkg - which determines a package name from the directory structure
tmpl - which generates the local index.html file from the
index.html.tmpl tempalte in the test directory.

At present any error in these packages results in t.Fatal being called.

The current set of refactored wasm tests, test-001-simple,
test-002-click, test-003-prop, test-004-component have all be refactoed
to use the new local testing package.

Also added a local .gitignore to each test directory, so that the
generated *.wasm, index.html and *_gen.go files are all ignored.
