+++ new-tests/test-002-click/.gitignore
@@ -1,0 +1,10 @@
+# Local .gitignore for the wasm test suite
+# We want to ignore any file that the test generates
+# These are the:
+# *.wasm - generated wasm files
+# *_gen.go - generated by vugugen via go generate
+# index.html - as thisis created dynamiclly from the index.html.tmpl template buy the test itself
+
+*.wasm
+*_gen.go
+index.html
