+++ new-tests/test-004-component/go.mod
@@ -5,13 +5,13 @@ replace github.com/vugu/vugu => ../..
 go 1.22.3
 
 require (
-	github.com/chromedp/cdproto v0.0.0-20240501202034-ef67d660e9fd
 	github.com/chromedp/chromedp v0.9.5
 	github.com/vugu/vjson v0.0.0-20200505061711-f9cbed27d3d9
 	github.com/vugu/vugu v0.0.0-00010101000000-000000000000
 )
 
 require (
+	github.com/chromedp/cdproto v0.0.0-20240501202034-ef67d660e9fd // indirect
 	github.com/chromedp/sysutil v1.0.0 // indirect
 	github.com/gobwas/httphead v0.1.0 // indirect
 	github.com/gobwas/pool v0.2.1 // indirect
